This commit adds a migration task; includes migrated data
This adds a migration task in a new folder, tasks, adds the task to the `scripts` property in package JSON, and includes the output of the script which is modified JSON files in /data.

There was one new file “preliminary-flood-insurance-rate” that exists because of a filename/id mismatch. This discrepancy should be flagged in the validations.